+++ packages/components/jest.config.ts
@@ -1,8 +1,15 @@
 import { jestConfig } from '@ttoss/config';
 
 const config = jestConfig({
-  setupFilesAfterEnv: ['./jest.setup.tsx'],
+  setupFilesAfterEnv: ['./tests/setupTests.tsx'],
   testEnvironment: 'jsdom',
+  moduleNameMapper: {
+    'react-markdown': [
+      '<rootDir>/../../node_modules/react-markdown/react-markdown.min.js',
+      '<rootDir>/node_modules/react-markdown/react-markdown.min.js',
+    ],
+  },
 });
 
+// eslint-disable-next-line import/no-default-export
 export default config;