+++ packages/components/src/index.ts
@@ -2,3 +2,4 @@ export * from './components/Accordion';
 export * from './components/InstallPwa';
 export * from './components/Modal';
 export * from './components/Toast';
+export * from './components/Markdown';