+++ docs/storybook/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.39.3](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.39.2...@docs/storybook@1.39.3) (2023-08-19)
+
+**Note:** Version bump only for package @docs/storybook
+
+## [1.39.2](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.39.1...@docs/storybook@1.39.2) (2023-08-16)
+
+**Note:** Version bump only for package @docs/storybook
+
+## [1.39.1](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.39.0...@docs/storybook@1.39.1) (2023-08-16)
+
+### Bug Fixes
+
+- carlin generate env ([#406](https://github.com/ttoss/ttoss/issues/406)) ([153ba71](https://github.com/ttoss/ttoss/commit/153ba71643461cdae076d3ba5779655f4988232c))
+
+# [1.39.0](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.38.18...@docs/storybook@1.39.0) (2023-08-15)
+
+### Features
+
+- create ttoss react-icons package ([#404](https://github.com/ttoss/ttoss/issues/404)) ([cd4c990](https://github.com/ttoss/ttoss/commit/cd4c990743da9bfd9d243d84adc38ad778824cc5))
+
+## [1.38.18](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.38.17...@docs/storybook@1.38.18) (2023-08-15)
+
+**Note:** Version bump only for package @docs/storybook
+
+## [1.38.17](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.38.16...@docs/storybook@1.38.17) (2023-08-11)
+
+**Note:** Version bump only for package @docs/storybook
+
 ## [1.38.16](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.38.15...@docs/storybook@1.38.16) (2023-08-03)
 
 **Note:** Version bump only for package @docs/storybook