+++ packages/react-i18n/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.24.0](https://github.com/ttoss/ttoss/compare/@ttoss/react-i18n@1.23.6...@ttoss/react-i18n@1.24.0) (2023-08-16)
+
+### Bug Fixes
+
+- carlin generate env ([#406](https://github.com/ttoss/ttoss/issues/406)) ([153ba71](https://github.com/ttoss/ttoss/commit/153ba71643461cdae076d3ba5779655f4988232c))
+
+### Features
+
+- create ttoss react-icons package ([#404](https://github.com/ttoss/ttoss/issues/404)) ([cd4c990](https://github.com/ttoss/ttoss/commit/cd4c990743da9bfd9d243d84adc38ad778824cc5))
+
 ## [1.23.6](https://github.com/ttoss/ttoss/compare/@ttoss/react-i18n@1.23.5...@ttoss/react-i18n@1.23.6) (2023-07-13)
 
 **Note:** Version bump only for package @ttoss/react-i18n