+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.6.0...@ttoss/theme@1.6.1) (2023-08-16)
+
+### Bug Fixes
+
+- carlin generate env ([#406](https://github.com/ttoss/ttoss/issues/406)) ([153ba71](https://github.com/ttoss/ttoss/commit/153ba71643461cdae076d3ba5779655f4988232c))
+
+# [1.6.0](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.5.6...@ttoss/theme@1.6.0) (2023-08-15)
+
+### Features
+
+- create ttoss react-icons package ([#404](https://github.com/ttoss/ttoss/issues/404)) ([cd4c990](https://github.com/ttoss/ttoss/commit/cd4c990743da9bfd9d243d84adc38ad778824cc5))
+
+## [1.5.6](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.5.5...@ttoss/theme@1.5.6) (2023-08-15)
+
+**Note:** Version bump only for package @ttoss/theme
+
 ## [1.5.5](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.5.4...@ttoss/theme@1.5.5) (2023-07-13)
 
 **Note:** Version bump only for package @ttoss/theme