+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.1](https://github.com/ttoss/ttoss/compare/@ttoss/ui@3.1.0...@ttoss/ui@3.1.1) (2023-08-16)
+
+### Bug Fixes
+
+- carlin generate env ([#406](https://github.com/ttoss/ttoss/issues/406)) ([153ba71](https://github.com/ttoss/ttoss/commit/153ba71643461cdae076d3ba5779655f4988232c))
+
+# [3.1.0](https://github.com/ttoss/ttoss/compare/@ttoss/ui@2.0.5...@ttoss/ui@3.1.0) (2023-08-15)
+
+### Features
+
+- create ttoss react-icons package ([#404](https://github.com/ttoss/ttoss/issues/404)) ([cd4c990](https://github.com/ttoss/ttoss/commit/cd4c990743da9bfd9d243d84adc38ad778824cc5))
+
+## [2.0.5](https://github.com/ttoss/ttoss/compare/@ttoss/ui@2.0.4...@ttoss/ui@2.0.5) (2023-08-15)
+
+**Note:** Version bump only for package @ttoss/ui
+
 ## [2.0.4](https://github.com/ttoss/ttoss/compare/@ttoss/ui@2.0.3...@ttoss/ui@2.0.4) (2023-07-17)
 
 **Note:** Version bump only for package @ttoss/ui