+++ terezinha-farm/iam/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.36](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.35...@terezinha-farm/iam@0.2.36) (2023-08-16)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
+## [0.2.35](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.34...@terezinha-farm/iam@0.2.35) (2023-08-11)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
 ## [0.2.34](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.33...@terezinha-farm/iam@0.2.34) (2023-08-03)
 
 **Note:** Version bump only for package @terezinha-farm/iam